requirements.txt
streamlit==1.56.0
paho-mqtt==2.1.0
plotly==6.1.2
pandas==2.2.3
pymodbus==3.13.0
